==> 04_metabolic_pathway_viz.Rmd <==
---
title: "04_metabolic_pathway_viz"
author: "Jagadeesh Puvvula"
date: "2025-01-02"
output: pdf_document
---

#needs R4.3
```{r}
library(pacman)
p_load(tidyverse, ggkegg, ggfx, ggraph, igraph, httr,
       clusterProfiler, dplyr, tidygraph)
```

#filtering unique molecules to highlight on the metabolic pathways
```{r}
res_df<- read_csv("~/Documents/air_polln_metabolomics/result/lm_mwas/placenta_mwas_anno.csv")

#summary of results
summary_df <- res_df |>
  filter(beta_dir %in% c("negative-significant", "positive-significant")) |>
  filter(sub_pathway =="Methionine, Cysteine, SAM and Taurine Metabolism") |>
  filter(startsWith(variable, "pm")) |>
  drop_na(kegg) |>
  select(c(10:13)) |>
  distinct()

#get the significant nodes to highlight
sig_nodes_x <- unique(unlist(sapply(summary_df$kegg, function(x) {
  # Split by comma and filter for values starting with 'C'
  values <- unlist(strsplit(x, ","))
  # Keep only those starting with 'C'
  values[grep("^C", values)]
})))

# Add 'cpd:' prefix
sig_nodes <- paste0("cpd:", sig_nodes_x)
```

#map compounds to module names in KEGG database
```{r}
get_module_ids <- function(compound_ids) {
  base_url <- "http://rest.kegg.jp/link/module/"
  
  results <- map_dfr(compound_ids, function(compound_id) {
    query_id <- if(!startsWith(compound_id, "cpd:")) {
      paste0("cpd:", compound_id)
    } else {
      compound_id
    }
    
    response <- GET(paste0(base_url, query_id))
    
    if (status_code(response) == 200) {
      content <- rawToChar(response$content)
      modules <- if(nchar(content) > 0) {
        paste(gsub(".*\t", "", strsplit(content, "\n")[[1]]), collapse = ", ")
      } else {
        "No modules found"
      }
    } else {
      modules <- "Error in query"
    }
    
    Sys.sleep(0.1)
    
    data.frame(
      Compound_ID = gsub("cpd:", "", query_id),
      Module_IDs = modules,
      stringsAsFactors = FALSE
    )
  })
  
  return(results)
}

#getting results
results <- get_module_ids(sig_nodes_x)

#
unique_modules <- unique(gsub("md:", "", unlist(strsplit(results$Module_IDs[results$Module_IDs != "No modules found"], ", "))))
cat(paste(sprintf('"%s"', unique_modules), collapse = ", "))
```


```{r}
g <- pathway("ko01100") |> 
  mutate(mod=
           highlight_set_nodes(sig_nodes, how="all"))

ggraph(g, layout="manual", x=x, y=y)+
    geom_node_rect(fill="grey85",
                   aes(filter=type == "ortholog"))+
    overlay_raw_map("ko01100")+
    geom_node_point(aes(filter=type == "compound"),
                    shape=21, fill="gray100", 
                    color="gray100", size=1)+
    ggfx::with_outer_glow(
      geom_node_point(aes(filter=mod, x=x, y=y), 
                      color="red",size=1),
      colour="yellow",expand=2
    )+
    theme_void() 
```

#function update
```{r}
plot_pathway <- function(pathway_id, cpd_list, module_list) {
    # Process pathway and highlight modules
    g <- pathway(pathway_id) |>
        process_line() |>
        mutate(is_highlight_cpd = name %in% cpd_list)
    
    # Add module highlights
    for (mod in module_list) {
        g <- g |> highlight_module(module(mod))
    }
    
    # Create the plot
    ggraph(g, x = x, y = y) +
        
        # Highlighted module edges
        ggfx::with_outer_glow(
            geom_edge_link0(width = 1,
                            aes(color = I(fgcolor),
                                filter = (M00017|M00034|M00035|M00368|M00609|M00953|M00021|M00338|M00672|M00673))),
            colour = "red", 
            expand = 0.25
        )  +
        # Highlighted compounds
        ggfx::with_outer_glow(
            geom_node_point(aes(filter = is_highlight_cpd),  
                            color = "red",
                            size = 4),
            colour = "orange",
            expand = 0.5
        )+
        # Base nodes
        geom_node_point(size = 0.25, 
                        aes(color = I(fgcolor),
                            filter = fgcolor != "none" & type != "line")) +
        # Base edges
        geom_edge_link0(width = 0.05, 
                        aes(color = I(fgcolor),
                            filter = type == "line" & fgcolor != "none")) +
        # Labels
        geom_node_text(size = 2,
                       aes(x = x, y = y,
                           label = graphics_name,
                           filter = name == "path:ko00270"),
                       repel = TRUE, 
                       family = "sans", 
                       bg.colour = "white") +
        theme_void()
}
```


```{r}
# Usage example:
plot_pathway("ko01100", sig_nodes, module_list=c("M00017", "M00034", "M00035", "M00368", "M00609", "M00953", "M00021", "M00338", "M00672", "M00673"))
```


```{r}
ggsave("~/Documents/air_polln_metabolomics/result/lm_mwas/mwas_metabol_path_map.tiff",
       dpi=300,
       bg="white",
       width = 10,
       height = 8)
```

#HMDB plotting
```{r}
#devtools::install_github("lanagarmire/lilikoi")
#library(lilikoi)
convertResults <- lilikoi.MetaTOpathway(hmdb_list, id.type="hmdb")
head(convertResults$table)
PDSmatrix <- lilikoi.PDSfun(convertResults$table)
```

```{r}
# Initialize RaMP database
rampDB <- RaMP()

# Prepare your HMDB list as a data frame with prefix column name
hmdb_list <- c("HMDB0000192", "HMDB0002757", "HMDB0000696", "HMDB0000574", 
               "HMDB0011745", "HMDB0002005", "HMDB0000731", "HMDB0001087", 
               "HMDB0000996", "HMDB0240253")

# Call getPathwayFromAnalyte() function (note: it requires 'conpass' if using local MySQL, but works with local rampDB by default)
pathways_df <- getPathwayFromAnalyte(analytes = hmdb_list,
                                    NameOrIds = "ids",
                                    analyte_type = "metabolites",
                                    find_synonym = FALSE,
                                    db = rampDB)

print(pathways_df)
```
